A cleaner format for printing GH status messages
Each PR is printed on a line in rich text. Thus the PR number can be
"linked" to the actual PR on GH. The status message block that
comprises of the PR name, from + to branch names, status of the
possible merge will be shown in a code block with easy-to-figure out
formatting.