Split `status all` into two commands
The old `status all` is split into two commands since the functionality
of the command is two unrelated tasks. A better idea would be to return
a summary of all open PRs when `status all` command is run. This change
achieves this. These are the two commands and changed functionality:

* `status all`

   This command will return a summary of number of open PRs, PRs open
   grouped by user, number of mergeable PRs and unmergeable PRs.

* `status conflicts`

   This command will return a semi-detailed information of each of the
   PRs that are not mergeable. The output will contain the URL of the
   PR, title of the PR, details of user and assignee.