Add more meaningful messages if no info
In case there are no open PRs on the repo, or if the user doesn't have
any open PRs, the message should be more meaningful than a bunch of
zeroes all over the place.

This change adds a check to see if the data that needs to be shown is
blank. If it is, it defaults to a simple message suggesting that there
are no open PRs. If it's not, the default stats are shown like before.