Use hubot's emit, on events and cleanup code
This makes it easier to cater to issues that were fixed by efa5534e33b
by in a more cleaner manner.

The change in `resp.send` is that the hubot Slack driver doesn't work if
we don't specify a room since there is no context in the case of an
event caught by `.on`. In the earlier case, that was implicit since
`hear` already has the current room in it's context.