Add GH hook for receiving PR events
For now, after every PR event that got received, we simply check for new
conflicts. This is a little inefficient and cumbersome. Ideally, the
message should be more detailed:

If there are *new* conflicts that did not exist up till that point,
those will be counted as conflicts-created-because-of-a-merge. Other
conflicts can just be mentioned though.