Add stricter command parsing regex
The commands can be given in three ways: 1. just the `status <cmd>` 2.
prepending `status <cmd>` with either `bot` or `@bot`. Ideally, the
bot has to listen to these three variations to avoid surprises:

status <cmd>
bot status <cmd>
@bot status <cmd>

Appending or prefixing these three variations with any other set of
words should not be considered as a command. This is a more robust
variation of the matcher regex.

Test for this regex is present at http://rubular.com/r/ZIZsNV1J6U

Problem:

Currently the bot name ("bot") is hardcoded into the regex. This should
be fixed.