Remove command descriptions in `bot status help`
Redirect the user to `bot help` instead.

This was added to cater to the case where a user might run the help
command like `bot status help` to know what available commands are
present. Now that the commands are documented on the main `@bot` command
level, these messages are not useful.