Fix issue of premature instantiation
If the CoffeeScript class files are stored in `scripts` directory, the
robot will try to instantiate the class files by passing the `robot`
object as an argument to their constructors. I've tried changing the
script in `index.coffee` to load just the `hubot-pr-status.coffee` file,
but they still got instantiated.

By moving the files to `src` folder, we can isolate the instantiation of
these classes only via code (by calling `new`). This fixes the warnings
generated by Github library because of invalid URLs.